requirements.txt
openai
python-dotenv
streamlit
pydantic
pillow 